+++ app/components/Editor/index.js
@@ -1,460 +1,579 @@
-import React from 'react';
+// @flow
 
-import beforeInput, { StringToTypeMap } from './util/beforeinput';
-import keyBindingFn from './util/keybinding';
-import rendererFn from './components/customrenderer';
-import createEditorState from './model/content';
+import React, { Component } from 'react';
 import {
+  Editor,
   EditorState,
-  convertToRaw,
-  convertFromRaw,
-  KeyBindingUtil,
-  Modifier,
-  AtomicBlockUtils
+  RichUtils,
+  SelectionState,
+  ContentBlock,
+  genKey,
+  Modifier
 } from 'draft-js';
-import { Block, HANDLED, NOT_HANDLED } from './util/constants';
+import isSoftNewlineEvent from 'draft-js/lib/isSoftNewlineEvent';
+import { OrderedMap } from 'immutable';
 
-import 'draft-js/dist/Draft.css';
-import { getCurrentBlock, addNewBlockAt } from './model';
+import AddButton from './components/addbutton';
+import Toolbar, { BLOCK_BUTTONS, INLINE_BUTTONS } from './components/toolbar';
+import LinkEditComponent from './components/LinkEditComponent';
 
-import ImageSideButton from './components/sides/image';
-
-import Editor from './editor';
-
+import rendererFn from './components/customrenderer';
+import RenderMap from './util/rendermap';
+import keyBindingFn from './util/keybinding';
 import {
-  setRenderOptions,
-  blockToHTML,
-  entityToHTML,
-  styleToHTML
-} from './exporter';
+  Block,
+  Entity as E,
+  HANDLED,
+  NOT_HANDLED,
+  KEY_COMMANDS
+} from './util/constants';
+import beforeInput, { StringToTypeMap } from './util/beforeinput';
+import blockStyleFn from './util/blockStyleFn';
+import {
+  getCurrentBlock,
+  resetBlockWithType,
+  addNewBlockAt,
+  isCursorBetweenLink
+} from './model';
 
-const newTypeMap = StringToTypeMap;
-newTypeMap['2.'] = Block.OL;
+import ImageButton from './components/sides/image';
 
-const { hasCommandModifier } = KeyBindingUtil;
-
-/*
-A demo for example editor. (Feature not built into medium-draft as too specific.)
-Convert quotes to curly quotes.
-*/
-const DQUOTE_START = '“';
-const DQUOTE_END = '”';
-const SQUOTE_START = '‘';
-const SQUOTE_END = '’';
-
-const newBlockToHTML = block => {
-  const blockType = block.type;
-  if (block.type === Block.ATOMIC) {
-    if (block.text === 'E') {
-      return {
-        start: '<figure class="md-block-atomic md-block-atomic-embed">',
-        end: '</figure>'
-      };
-    } else if (block.text === '-') {
-      return (
-        <div className="md-block-atomic md-block-atomic-break">
-          <hr />
-        </div>
-      );
-    }
-  }
-  return blockToHTML(block);
+type State = {
+  editorState: EditorState
 };
 
-const newEntityToHTML = (entity, originalText) => {
-  if (entity.type === 'embed') {
-    return (
-      <div>
-        <a
-          className="embedly-card"
-          href={entity.data.url}
-          data-card-controls="0"
-          data-card-theme="dark"
-        >
-          Embedded ― {entity.data.url}
-        </a>
-      </div>
-    );
-  }
-  return entityToHTML(entity, originalText);
+type Props = {
+  placeholder?: string
 };
 
-const handleBeforeInput = (editorState, str, onChange) => {
-  if (str === '"' || str === "'") {
-    const currentBlock = getCurrentBlock(editorState);
-    const selectionState = editorState.getSelection();
-    const contentState = editorState.getCurrentContent();
-    const text = currentBlock.getText();
-    const len = text.length;
-    if (selectionState.getAnchorOffset() === 0) {
-      onChange(
-        EditorState.push(
-          editorState,
-          Modifier.insertText(
-            contentState,
-            selectionState,
-            str === '"' ? DQUOTE_START : SQUOTE_START
-          ),
-          'transpose-characters'
-        )
-      );
-      return HANDLED;
-    } else if (len > 0) {
-      const lastChar = text[len - 1];
-      if (lastChar !== ' ') {
-        onChange(
-          EditorState.push(
-            editorState,
-            Modifier.insertText(
-              contentState,
-              selectionState,
-              str === '"' ? DQUOTE_END : SQUOTE_END
-            ),
-            'transpose-characters'
-          )
-        );
-      } else {
-        onChange(
-          EditorState.push(
-            editorState,
-            Modifier.insertText(
-              contentState,
-              selectionState,
-              str === '"' ? DQUOTE_START : SQUOTE_START
-            ),
-            'transpose-characters'
-          )
-        );
-      }
-      return HANDLED;
-    }
-  }
-  return beforeInput(editorState, str, onChange, newTypeMap);
-};
+export default class CustomEditor extends Component<Props, State> {
+  props: Props;
 
-class SeparatorSideButton extends React.Component {
-  constructor(props) {
-    super(props);
-    this.onClick = this.onClick.bind(this);
-  }
-
-  onClick() {
-    let editorState = this.props.getEditorState();
-    const content = editorState.getCurrentContent();
-    const contentWithEntity = content.createEntity(
-      'separator',
-      'IMMUTABLE',
-      {}
-    );
-    const entityKey = contentWithEntity.getLastCreatedEntityKey();
-    editorState = EditorState.push(
-      editorState,
-      contentWithEntity,
-      'create-entity'
-    );
-    this.props.setEditorState(
-      AtomicBlockUtils.insertAtomicBlock(editorState, entityKey, '-')
-    );
-    this.props.close();
-  }
-
-  render() {
-    return (
-      <button
-        className="md-sb-button md-sb-img-button"
-        type="button"
-        title="Add a separator"
-        onClick={this.onClick}
-      >
-        <i className="fa fa-minus" />
-      </button>
-    );
-  }
-}
-
-const AtomicSeparatorComponent = props => <hr />;
-
-const AtomicBlock = props => {
-  const { blockProps, block } = props;
-  const content = blockProps.getEditorState().getCurrentContent();
-  const entity = content.getEntity(block.getEntityAt(0));
-  const data = entity.getData();
-  const type = entity.getType();
-  if (blockProps.components[type]) {
-    const AtComponent = blockProps.components[type];
-    return (
-      <div
-        className={`md-block-atomic-wrapper md-block-atomic-wrapper-${type}`}
-      >
-        <AtComponent data={data} />
-      </div>
-    );
-  }
-  return (
-    <p>
-      Block of type <b>{type}</b> is not supported.
-    </p>
-  );
-};
-
-export default class App extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      editorState: createEditorState(),
-      editorEnabled: true,
-      placeholder: 'Write here...'
-    };
-
-    this.onChange = (editorState, callback = null) => {
-      if (this.state.editorEnabled) {
-        this.setState({ editorState }, () => {
-          if (callback) {
-            callback();
-          }
-        });
-      }
-    };
-
-    this.sideButtons = [
+  static defaultProps = {
+    beforeInput,
+    keyBindingFn,
+    blockStyleFn,
+    rendererFn,
+    editorEnabled: true,
+    spellCheck: true,
+    stringToTypeMap: StringToTypeMap,
+    blockRenderMap: RenderMap,
+    blockButtons: BLOCK_BUTTONS,
+    inlineButtons: INLINE_BUTTONS,
+    placeholder: 'Write your story...',
+    continuousBlocks: [
+      Block.UNSTYLED,
+      Block.BLOCKQUOTE,
+      Block.OL,
+      Block.UL,
+      Block.CODE,
+      Block.TODO
+    ],
+    sideButtons: [
       {
         title: 'Image',
-        component: ImageSideButton
-      },
-      {
-        title: 'Separator',
-        component: SeparatorSideButton
+        component: ImageButton
       }
-    ];
+    ],
+    disableToolbar: false,
+    showLinkEditToolbar: true,
+    toolbarConfig: {}
+  };
 
-    this.exporter = setRenderOptions({
-      styleToHTML,
-      blockToHTML: newBlockToHTML,
-      entityToHTML: newEntityToHTML
-    });
+  state: State = {
+    editorState: null
+  };
 
-    this.getEditorState = () => this.state.editorState;
+  focus = () => this._editorNode.focus();
+  onChange = (editorState, cb) => {
+    this.props.onChange(editorState, cb);
+  };
 
-    this.logData = this.logData.bind(this);
-    this.renderHTML = this.renderHTML.bind(this);
-    this.toggleEdit = this.toggleEdit.bind(this);
-    this.fetchData = this.fetchData.bind(this);
-    this.loadSavedData = this.loadSavedData.bind(this);
-    this.keyBinding = this.keyBinding.bind(this);
-    this.handleKeyCommand = this.handleKeyCommand.bind(this);
-    this.handleDroppedFiles = this.handleDroppedFiles.bind(this);
-    this.handleReturn = this.handleReturn.bind(this);
-  }
+  getEditorState = () => this.state.editorState;
 
-  componentDidMount() {
-    setTimeout(this.fetchData, 1000);
-  }
+  blockRendererFn = this.props.rendererFn(this.onChange, this.getEditorState);
 
-  rendererFn(setEditorState, getEditorState) {
-    const atomicRenderers = {
-      separator: AtomicSeparatorComponent
-    };
-    const rFnOld = rendererFn(setEditorState, getEditorState);
-    const rFnNew = contentBlock => {
-      const type = contentBlock.getType();
-      switch (type) {
-        case Block.ATOMIC:
-          return {
-            component: AtomicBlock,
-            editable: false,
-            props: {
-              components: atomicRenderers,
-              getEditorState
-            }
-          };
-        default:
-          return rFnOld(contentBlock);
-      }
-    };
-    return rFnNew;
-  }
-
-  keyBinding(e) {
-    if (hasCommandModifier(e)) {
-      if (e.which === 83) {
-        /* Key S */
-        return 'editor-save';
-      }
-      // else if (e.which === 74 /* Key J */) {
-      //  return 'do-nothing';
-      //}
+  /**
+   * Implemented to provide nesting of upto 2 levels in ULs or OLs.
+   */
+  onTab(e) {
+    const { editorState } = this.props;
+    const newEditorState = RichUtils.onTab(e, editorState, 2);
+    if (newEditorState !== editorState) {
+      this.onChange(newEditorState);
     }
-    if (e.altKey === true) {
-      if (e.shiftKey === true) {
-        switch (e.which) {
-          /* Alt + Shift + L */
-          case 76:
-            return 'load-saved-data';
-          /* Key E */
-          // case 69: return 'toggle-edit-mode';
+  }
+
+  onUpArrow = e => {
+    const { editorState } = this.props;
+    const content = editorState.getCurrentContent();
+    const selection = editorState.getSelection();
+    const key = selection.getAnchorKey();
+    const currentBlock = content.getBlockForKey(key);
+    const firstBlock = content.getFirstBlock();
+    if (firstBlock.getKey() === key) {
+      if (firstBlock.getType().indexOf(Block.ATOMIC) === 0) {
+        e.preventDefault();
+        const newBlock = new ContentBlock({
+          type: Block.UNSTYLED,
+          key: genKey()
+        });
+        const newBlockMap = OrderedMap([[newBlock.getKey(), newBlock]]).concat(
+          content.getBlockMap()
+        );
+        const newContent = content.merge({
+          blockMap: newBlockMap,
+          selectionAfter: selection.merge({
+            anchorKey: newBlock.getKey(),
+            focusKey: newBlock.getKey(),
+            anchorOffset: 0,
+            focusOffset: 0,
+            isBackward: false
+          })
+        });
+        this.onChange(
+          EditorState.push(editorState, newContent, 'insert-characters')
+        );
+      }
+    } else if (currentBlock.getType().indexOf(Block.ATOMIC) === 0) {
+      const blockBefore = content.getBlockBefore(key);
+      if (!blockBefore) {
+        return;
+      }
+      e.preventDefault();
+      const newSelection = selection.merge({
+        anchorKey: blockBefore.getKey(),
+        focusKey: blockBefore.getKey(),
+        anchorOffset: blockBefore.getLength(),
+        focusOffset: blockBefore.getLength(),
+        isBackward: false
+      });
+      this.onChange(EditorState.forceSelection(editorState, newSelection));
+    }
+  };
+
+  /*
+  Adds a hyperlink on the selected text with some basic checks.
+  */
+  setLink(url) {
+    let { editorState } = this.props;
+    const selection = editorState.getSelection();
+    const content = editorState.getCurrentContent();
+    let entityKey = null;
+    let newUrl = url;
+    if (url !== '') {
+      if (url.indexOf('http') === -1) {
+        if (url.indexOf('@') >= 0) {
+          newUrl = `mailto:${newUrl}`;
+        } else {
+          newUrl = `http://${newUrl}`;
         }
       }
-      if (e.which === 72 /* Key H */) {
-        return 'toggleinline:HIGHLIGHT';
-      }
+      const contentWithEntity = content.createEntity(E.LINK, 'MUTABLE', {
+        url: newUrl
+      });
+      editorState = EditorState.push(
+        editorState,
+        contentWithEntity,
+        'create-entity'
+      );
+      entityKey = contentWithEntity.getLastCreatedEntityKey();
     }
-    return keyBindingFn(e);
+    this.onChange(
+      RichUtils.toggleLink(editorState, selection, entityKey),
+      this.focus
+    );
   }
 
+  /**
+   * Override which text modifications are available according BLOCK_BUTTONS style property.
+   * Defaults all of them if no toolbarConfig.block passed:
+   *   block: ['ordered-list-item', 'unordered-list-item', 'blockquote', 'header-three', 'todo'],
+   * Example parameter: toolbarConfig = {
+   *   block: ['ordered-list-item', 'unordered-list-item'],
+   *   inline: ['BOLD', 'ITALIC', 'UNDERLINE', 'hyperlink'],
+   * };
+   */
+  configureToolbarBlockOptions(toolbarConfig) {
+    return toolbarConfig && toolbarConfig.block
+      ? toolbarConfig.block
+          .map(type => BLOCK_BUTTONS.find(button => button.style === type))
+          .filter(button => button !== undefined)
+      : this.props.blockButtons;
+  }
+
+  /**
+   * Override which text modifications are available according INLINE_BUTTONS style property.
+   * CASE SENSITIVE. Would be good clean up to lowercase inline styles consistently.
+   * Defaults all of them if no toolbarConfig.inline passed:
+   *   inline: ['BOLD', 'ITALIC', 'UNDERLINE', 'hyperlink', 'HIGHLIGHT'],
+   * Example parameter: toolbarConfig = {
+   *   block: ['ordered-list-item', 'unordered-list-item'],
+   *   inline: ['BOLD', 'ITALIC', 'UNDERLINE', 'hyperlink'],
+   * };
+   */
+  configureToolbarInlineOptions(toolbarConfig) {
+    return toolbarConfig && toolbarConfig.inline
+      ? toolbarConfig.inline
+          .map(type => INLINE_BUTTONS.find(button => button.style === type))
+          .filter(button => button !== undefined)
+      : this.props.inlineButtons;
+  }
+
+  /*
+  Handles custom commands based on various key combinations. First checks
+  for some built-in commands. If found, that command's function is apllied and returns.
+  If not found, it checks whether parent component handles that command or not.
+  Some of the internal commands are:
+
+  - showlinkinput -> Opens up the link input tooltip if some text is selected.
+  - add-new-block -> Adds a new block at the current cursor position.
+  - changetype:block-type -> If the command starts with `changetype:` and
+    then succeeded by the block type, the current block will be converted to that particular type.
+  - toggleinline:inline-type -> If the command starts with `toggleinline:` and
+    then succeeded by the inline type, the current selection's inline type will be
+    togglled.
+  */
   handleKeyCommand(command) {
-    if (command === 'editor-save') {
-      window.localStorage['editor'] = JSON.stringify(
-        convertToRaw(this.state.editorState.getCurrentContent())
-      );
-      window.ga('send', 'event', 'draftjs', command);
-      return true;
-    } else if (command === 'load-saved-data') {
-      this.loadSavedData();
-      return true;
-    } else if (command === 'toggle-edit-mode') {
-      this.toggleEdit();
-    }
-    return false;
-  }
-
-  fetchData() {
-    window.ga('send', 'event', 'draftjs', 'load-data', 'ajax');
-    this.setState({
-      placeholder: 'Loading...'
-    });
-    const req = new XMLHttpRequest();
-    req.open('GET', 'data.json', true);
-    req.onreadystatechange = () => {
-      if (req.readyState === 4) {
-        const data = JSON.parse(req.responseText);
-        this.setState(
-          {
-            editorState: createEditorState(data),
-            placeholder: 'Write here...'
-          },
-          () => {
-            this._editor.focus();
-          }
-        );
-        window.ga('send', 'event', 'draftjs', 'data-success');
+    // console.log(command);
+    const { editorState } = this.props;
+    if (this.props.handleKeyCommand) {
+      const behaviour = this.props.handleKeyCommand(command);
+      if (behaviour === HANDLED || behaviour === true) {
+        return HANDLED;
       }
-    };
-    req.send();
-  }
-
-  logData(e) {
-    const currentContent = this.state.editorState.getCurrentContent();
-    const es = convertToRaw(currentContent);
-    console.log(es);
-    console.log(this.state.editorState.getSelection().toJS());
-    window.ga('send', 'event', 'draftjs', 'log-data');
-  }
-
-  renderHTML(e) {
-    const currentContent = this.state.editorState.getCurrentContent();
-    const eHTML = this.exporter(currentContent);
-    var newWin = window.open(
-      `${window.location.pathname}rendered.html`,
-      'windowName',
-      `height=${window.screen.height},width=${window.screen.wdith}`
-    );
-    newWin.onload = () => newWin.postMessage(eHTML, window.location.origin);
-  }
-
-  loadSavedData() {
-    const data = window.localStorage.getItem('editor');
-    if (data === null) {
-      return;
     }
-    try {
-      const blockData = JSON.parse(data);
-      console.log(blockData);
-      this.onChange(
-        EditorState.push(this.state.editorState, convertFromRaw(blockData)),
-        this._editor.focus
-      );
-    } catch (e) {
-      console.log(e);
-    }
-    window.ga('send', 'event', 'draftjs', 'load-data', 'localstorage');
-  }
-
-  toggleEdit(e) {
-    this.setState(
-      {
-        editorEnabled: !this.state.editorEnabled
-      },
-      () => {
-        window.ga(
-          'send',
-          'event',
-          'draftjs',
-          'toggle-edit',
-          this.state.editorEnabled + ''
-        );
+    if (command === KEY_COMMANDS.showLinkInput()) {
+      if (!this.props.disableToolbar && this.toolbar) {
+        // For some reason, scroll is jumping sometimes for the below code.
+        // Debug and fix it later.
+        const isCursorLink = isCursorBetweenLink(editorState);
+        if (isCursorLink) {
+          this.editLinkAfterSelection(
+            isCursorLink.blockKey,
+            isCursorLink.entityKey
+          );
+          return HANDLED;
+        }
+        this.toolbar.handleLinkInput(null, true);
+        return HANDLED;
       }
-    );
-  }
-
-  handleDroppedFiles(selection, files) {
-    window.ga(
-      'send',
-      'event',
-      'draftjs',
-      'filesdropped',
-      files.length + ' files'
-    );
-    const file = files[0];
-    if (file.type.indexOf('image/') === 0) {
-      // eslint-disable-next-line no-undef
-      const src = URL.createObjectURL(file);
-      this.onChange(
-        addNewBlockAt(
-          this.state.editorState,
-          selection.getAnchorKey(),
-          Block.IMAGE,
-          {
-            src
-          }
-        )
-      );
+      return NOT_HANDLED;
+    } else if (command === KEY_COMMANDS.unlink()) {
+      const isCursorLink = isCursorBetweenLink(editorState);
+      if (isCursorLink) {
+        this.removeLink(isCursorLink.blockKey, isCursorLink.entityKey);
+        return HANDLED;
+      }
+    }
+    /* else if (command === KEY_COMMANDS.addNewBlock()) {
+      const { editorState } = this.props;
+      this.onChange(addNewBlock(editorState, Block.BLOCKQUOTE));
+      return HANDLED;
+    } */
+    const block = getCurrentBlock(editorState);
+    const currentBlockType = block.getType();
+    // if (command === KEY_COMMANDS.deleteBlock()) {
+    //   if (currentBlockType.indexOf(Block.ATOMIC) === 0 && block.getText().length === 0) {
+    //     this.onChange(resetBlockWithType(editorState, Block.UNSTYLED, { text: '' }));
+    //     return HANDLED;
+    //   }
+    //   return NOT_HANDLED;
+    // }
+    if (command.indexOf(`${KEY_COMMANDS.changeType()}`) === 0) {
+      let newBlockType = command.split(':')[1];
+      // const currentBlockType = block.getType();
+      if (currentBlockType === Block.ATOMIC) {
+        return HANDLED;
+      }
+      if (
+        currentBlockType === Block.BLOCKQUOTE &&
+        newBlockType === Block.CAPTION
+      ) {
+        newBlockType = Block.BLOCKQUOTE_CAPTION;
+      } else if (
+        currentBlockType === Block.BLOCKQUOTE_CAPTION &&
+        newBlockType === Block.CAPTION
+      ) {
+        newBlockType = Block.BLOCKQUOTE;
+      }
+      this.onChange(RichUtils.toggleBlockType(editorState, newBlockType));
+      return HANDLED;
+    } else if (command.indexOf(`${KEY_COMMANDS.toggleInline()}`) === 0) {
+      const inline = command.split(':')[1];
+      this._toggleInlineStyle(inline);
+      return HANDLED;
+    }
+    const newState = RichUtils.handleKeyCommand(editorState, command);
+    if (newState) {
+      this.onChange(newState);
       return HANDLED;
     }
     return NOT_HANDLED;
   }
 
+  /*
+  This function is responsible for emitting various commands based on various key combos.
+  */
+  handleBeforeInput(str) {
+    return this.props.beforeInput(
+      this.props.editorState,
+      str,
+      this.onChange,
+      this.props.stringToTypeMap
+    );
+  }
+
+  /*
+  By default, it handles return key for inserting soft breaks (BRs in HTML) and
+  also instead of inserting a new empty block after current empty block, it first check
+  whether the current block is of a type other than `unstyled`. If yes, current block is
+  simply converted to an unstyled empty block. If RETURN is pressed on an unstyled block
+  default behavior is executed.
+  */
   handleReturn(e) {
-    // const currentBlock = getCurrentBlock(this.state.editorState);
-    // var text = currentBlock.getText();
+    if (this.props.handleReturn) {
+      const behavior = this.props.handleReturn();
+      if (behavior === HANDLED || behavior === true) {
+        return HANDLED;
+      }
+    }
+    const { editorState } = this.props;
+    if (isSoftNewlineEvent(e)) {
+      this.onChange(RichUtils.insertSoftNewline(editorState));
+      return HANDLED;
+    }
+    if (!e.altKey && !e.metaKey && !e.ctrlKey) {
+      const currentBlock = getCurrentBlock(editorState);
+      const blockType = currentBlock.getType();
+
+      if (blockType.indexOf(Block.ATOMIC) === 0) {
+        this.onChange(addNewBlockAt(editorState, currentBlock.getKey()));
+        return HANDLED;
+      }
+
+      if (currentBlock.getLength() === 0) {
+        switch (blockType) {
+          case Block.UL:
+          case Block.OL:
+          case Block.BLOCKQUOTE:
+          case Block.BLOCKQUOTE_CAPTION:
+          case Block.CAPTION:
+          case Block.TODO:
+          case Block.H2:
+          case Block.H3:
+          case Block.H1:
+            this.onChange(resetBlockWithType(editorState, Block.UNSTYLED));
+            return HANDLED;
+          default:
+            return NOT_HANDLED;
+        }
+      }
+
+      const selection = editorState.getSelection();
+
+      if (
+        selection.isCollapsed() &&
+        currentBlock.getLength() === selection.getStartOffset()
+      ) {
+        if (this.props.continuousBlocks.indexOf(blockType) < 0) {
+          this.onChange(addNewBlockAt(editorState, currentBlock.getKey()));
+          return HANDLED;
+        }
+        return NOT_HANDLED;
+      }
+      return NOT_HANDLED;
+    }
     return NOT_HANDLED;
   }
 
+  /*
+  The function documented in `draft-js` to be used to toggle block types (mainly
+  for some key combinations handled by default inside draft-js).
+  */
+  _toggleBlockType(blockType) {
+    const type = RichUtils.getCurrentBlockType(this.props.editorState);
+    if (type.indexOf(`${Block.ATOMIC}:`) === 0) {
+      return;
+    }
+    this.onChange(RichUtils.toggleBlockType(this.props.editorState, blockType));
+  }
+
+  /*
+  The function documented in `draft-js` to be used to toggle inline styles of selection (mainly
+  for some key combinations handled by default inside draft-js).
+  */
+  _toggleInlineStyle(inlineStyle) {
+    this.onChange(
+      RichUtils.toggleInlineStyle(this.props.editorState, inlineStyle)
+    );
+  }
+
+  removeLink = (blockKey, entityKey) => {
+    const { editorState } = this.props;
+    const content = editorState.getCurrentContent();
+    const block = content.getBlockForKey(blockKey);
+    const oldSelection = editorState.getSelection();
+    block.findEntityRanges(
+      character => {
+        const eKey = character.getEntity();
+        return eKey === entityKey;
+      },
+      (start, end) => {
+        const selection = new SelectionState({
+          anchorKey: blockKey,
+          focusKey: blockKey,
+          anchorOffset: start,
+          focusOffset: end
+        });
+        const newEditorState = EditorState.forceSelection(
+          RichUtils.toggleLink(editorState, selection, null),
+          oldSelection
+        );
+        this.onChange(newEditorState, this.focus);
+      }
+    );
+  };
+
+  editLinkAfterSelection = (blockKey, entityKey = null) => {
+    if (entityKey === null) {
+      return;
+    }
+    const { editorState } = this.props;
+    const content = editorState.getCurrentContent();
+    const block = content.getBlockForKey(blockKey);
+    block.findEntityRanges(
+      character => {
+        const eKey = character.getEntity();
+        return eKey === entityKey;
+      },
+      (start, end) => {
+        const selection = new SelectionState({
+          anchorKey: blockKey,
+          focusKey: blockKey,
+          anchorOffset: start,
+          focusOffset: end
+        });
+        const newEditorState = EditorState.forceSelection(
+          editorState,
+          selection
+        );
+        this.onChange(newEditorState);
+        setTimeout(() => {
+          if (this.toolbar) {
+            this.toolbar.handleLinkInput(null, true);
+          }
+        }, 100);
+      }
+    );
+  };
+
+  /**
+   * Handle pasting when cursor is in an image block. Paste the text as the
+   * caption. Otherwise, let Draft do its thing.
+   */
+  handlePastedText = (text, html, es) => {
+    const currentBlock = getCurrentBlock(this.props.editorState);
+    if (currentBlock.getType() === Block.IMAGE) {
+      const { editorState } = this.props;
+      const content = editorState.getCurrentContent();
+      this.onChange(
+        EditorState.push(
+          editorState,
+          Modifier.insertText(content, editorState.getSelection(), text)
+        )
+      );
+      return HANDLED;
+    }
+    if (
+      this.props.handlePastedText &&
+      this.props.handlePastedText(text, html, es) === HANDLED
+    ) {
+      return HANDLED;
+    }
+    return NOT_HANDLED;
+  };
+
+  /*
+  Renders the `Editor`, `Toolbar` and the side `AddButton`.
+  */
   render() {
-    const { editorState, editorEnabled } = this.state;
+    const {
+      editorState,
+      editorEnabled,
+      disableToolbar,
+      showLinkEditToolbar,
+      toolbarConfig
+    } = this.props;
+    const showAddButton = editorEnabled;
+    const editorClass = `md-RichEditor-editor${!editorEnabled
+      ? ' md-RichEditor-readonly'
+      : ''}`;
+    let isCursorLink = false;
+    if (editorEnabled && showLinkEditToolbar) {
+      isCursorLink = isCursorBetweenLink(editorState);
+    }
+    const blockButtons = this.configureToolbarBlockOptions(toolbarConfig);
+    const inlineButtons = this.configureToolbarInlineOptions(toolbarConfig);
     return (
-      <div>
-        <div className="editor-action">
-          <button onClick={this.logData}>Log State</button>
-          <button onClick={this.renderHTML}>Render HTML</button>
-          <button onClick={this.toggleEdit}>Toggle Edit</button>
+      <div className="md-RichEditor-root">
+        <div className={editorClass}>
+          <Editor
+            ref={node => {
+              this._editorNode = node;
+            }}
+            {...this.props}
+            editorState={editorState}
+            blockRendererFn={this.blockRendererFn}
+            blockStyleFn={this.props.blockStyleFn}
+            onChange={this.onChange}
+            onTab={this.onTab}
+            onUpArrow={this.onUpArrow}
+            blockRenderMap={this.props.blockRenderMap}
+            handleKeyCommand={this.handleKeyCommand}
+            handleBeforeInput={this.handleBeforeInput}
+            handleReturn={this.handleReturn}
+            handlePastedText={this.handlePastedText}
+            readOnly={!editorEnabled}
+            keyBindingFn={this.props.keyBindingFn}
+            placeholder={this.props.placeholder}
+            spellCheck={editorEnabled && this.props.spellCheck}
+          />
+          {this.props.sideButtons.length > 0 &&
+            showAddButton && (
+              <AddButton
+                editorState={editorState}
+                getEditorState={this.getEditorState}
+                setEditorState={this.onChange}
+                focus={this.focus}
+                sideButtons={this.props.sideButtons}
+              />
+            )}
+          {!disableToolbar && (
+            <Toolbar
+              ref={c => {
+                this.toolbar = c;
+              }}
+              editorNode={this._editorNode}
+              editorState={editorState}
+              toggleBlockType={this.toggleBlockType}
+              toggleInlineStyle={this.toggleInlineStyle}
+              editorEnabled={editorEnabled}
+              setLink={this.setLink}
+              focus={this.focus}
+              blockButtons={blockButtons}
+              inlineButtons={inlineButtons}
+            />
+          )}
+          {isCursorLink && (
+            <LinkEditComponent
+              {...isCursorLink}
+              editorState={editorState}
+              removeLink={this.removeLink}
+              editLink={this.editLinkAfterSelection}
+            />
+          )}
         </div>
-        <Editor
-          ref={e => {
-            this._editor = e;
-          }}
-          editorState={editorState}
-          onChange={this.onChange}
-          editorEnabled={editorEnabled}
-          handleDroppedFiles={this.handleDroppedFiles}
-          handleKeyCommand={this.handleKeyCommand}
-          placeholder={this.state.placeholder}
-          keyBindingFn={this.keyBinding}
-          beforeInput={handleBeforeInput}
-          handleReturn={this.handleReturn}
-          sideButtons={this.sideButtons}
-          rendererFn={this.rendererFn}
-        />
       </div>
     );
   }
